database + images updated

+++ app/Http/Controllers/BottleSizeController.php
@@ -12,20 +12,20 @@ class BottleSizeController extends Controller
 {
     public function showBottleSizes()
     {
+        if(Cart::count() > 0) {
+            Cart::destroy();
+        }
         $bottles = BottleSize::all();
-
+        //return response()->json(['bottles' => $bottles ]);
         return view('steps/step1ChooseSize', compact('bottles'));
     }
 
     public function showInhalt(Request $request, $bottleID)
     { 
-        Cart::destroy();
-        
         $bottle = BottleSize::findOrFail($bottleID);
         $request->session()->put('bottle', $bottle);
-
+        //return response()->json([]);
         $zutaten = Ingrediente::where('type', IngredienteType::FRUITS)->get();
-      
         return view('steps/step2ChooseIngrediente', compact('zutaten'));
     }
